[videotexturebackend] Use persistent bindings of graphics buffers to texture ids. Contributes to JB#49430
Binding an EGL image to a texture ID can be expensive on some hardware.
So rather than using a single texture ID and changing the buffer it's
bound to on every update create a texture ID for each buffer and re-use
those as the buffers are recycled by the pipeline.

This requires more direct access to the gstreamer buffers in order to
associate an id to a buffer.